Circular buffer grow is now correct.

+++ src/extra-io/net/multiphasicapps/io/CircularByteBuffer.java
@@ -82,7 +82,7 @@ public class CircularByteBuffer
 	protected CircularGenericBuffer arrayWrite(byte[] __arr, int __dx,
 		Byte __v)
 	{
-		__arr[__dx] = __v;
+		__arr[__dx] = __v.byteValue();
 		return this;
 	}
